Moves ZebraScanRpcImpl into its own file
