+++ zebra-grpc/src/methods/types.rs
@@ -1,1 +1,3 @@
+//! Implements methods on RPC request/response types
+
 pub mod scan;
